improve class name
