+++ computer-core/src/main/java/com/baidu/hugegraph/computer/core/input/loader/LoaderFileGraphFetcher.java
@@ -26,15 +26,16 @@ import com.baidu.hugegraph.computer.core.input.InputSplit;
 import com.baidu.hugegraph.computer.core.input.VertexFetcher;
 import com.baidu.hugegraph.computer.core.rpc.InputSplitRpcService;
 
-public class LoaderGraphFetcher implements GraphFetcher {
+public class LoaderFileGraphFetcher implements GraphFetcher {
 
     private final InputSplitRpcService rpcService;
     private final VertexFetcher vertexFetcher;
     private final EdgeFetcher edgeFetcher;
 
-    public LoaderGraphFetcher(Config config, InputSplitRpcService rpcService) {
-        this.vertexFetcher = new LoaderVertxFetcher(config);
-        this.edgeFetcher = new LoaderEdgeFetcher(config);
+    public LoaderFileGraphFetcher(Config config,
+                                  InputSplitRpcService rpcService) {
+        this.vertexFetcher = new FileVertxFetcher(config);
+        this.edgeFetcher = new FileEdgeFetcher(config);
         this.rpcService = rpcService;
     }
